Lock the native buffer before accessing its contents.
Fix for #2.

Before the fix the JNA structures were present in the method signature. This
meant that those structures would be automatically read (which is usually
nice and convenient) but in this case the native memory was being accessed
before it was locked (and therefore there was nothing preventing the contents
of that memory from being changed while it was being accessed).

This caused a race and would lead to an intermittent hang, usually observed
when the native library seemingly stopped sending frames of picture data
(always seemed to be the last frame).

The fix therefore is to replace the JNA structure object in the method
signature with a simple Pointer, then to use the native API to lock that
pointer, and only then access the data in the object.