Throw a CaptureFailedException if there is an error during capture
This covers the case where the capture thread was interrupted while
waiting, or the (optional) timeout expired when waiting for the
native capture to complete.

If the exception is thrown, client applcations should close the
camera instance and not use it again.

The default behaviour is to capture without a timeout, as was the
case previously.