Use JNI rather than JNA
Removed all of the JNA binding classes, there is a companion project
at github.com/caprica/picam-native that provides a small native
library that is required by this project.

Removed the JNA and the logging dependencies.

Some small API changes, so picam-1.x code will not be a drop-in
replacement - but it will be close.